tools/devops.py
from .base import Tool, OperationResult
from .git_tools import GitTools

class DevOpsTools(Tool):
    name: str = "DevOps"
    description: str = "DevOps operations for deployment and environment management"
    
    def __init__(self):
        self.git_tools = GitTools()
    
    def push_changes(self, remote: str, branch: str) -> OperationResult:
        return self.git_tools._execute_git_command(['push', remote, branch])

    def create_pull_request(self, source: str, target: str) -> OperationResult:
        # TODO: Implement pull request creation
        # This might require integration with a specific Git hosting service API (e.g., GitHub, GitLab)
        return OperationResult(success=False, message="Pull request creation not implemented")

    def deploy_environment(self, env: str) -> OperationResult:
        # TODO: Implement environment deployment
        # This might involve running deployment scripts or using deployment tools
        return OperationResult(success=False, message="Environment deployment not implemented")

    def manage_remote_connection(self) -> OperationResult:
        # TODO: Implement remote connection management
        # This might involve setting up SSH connections or managing cloud service credentials
        return OperationResult(success=False, message="Remote connection management not implemented")
