tests/tools/test_devops.py
import pytest
from unittest.mock import patch, MagicMock
from tools.devops import DevOpsTools
import tempfile
import os

@pytest.fixture
def devops_tools():
    return DevOpsTools()

class TestDevOps:
    @pytest.fixture
    def dev_env(self):
        with tempfile.TemporaryDirectory() as tmpdirname:
            yield tmpdirname
            
    def test_push_operations(self, devops_tools):
        with patch.object(DevOpsTools, 'push_changes') as mock_push:
            mock_push.return_value = MagicMock(success=True)
            
            # 测试代码推送
            result = devops_tools.push_changes('origin', 'main')
            assert result.success

    def test_deploy_operations(self, dev_env):
        devops_tools = DevOpsTools()
        
        # 测试环境部署
        result = devops_tools.deploy_environment("test")
        assert result.success is False  # 因为这是未实现的功能
        assert "not implemented" in result.message.lower()

    def test_remote_operations(self, devops_tools):
        # 测试远程连接管理
        result = devops_tools.manage_remote_connection()
        assert not result.success  # 当前未实现
        assert "Remote connection management not implemented" in result.message
